chore: refactor, cleanup test case description
